requirements.txt
absl-py==2.1.0
alembic==1.14.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
configparser==7.1.0
DateTime==5.5
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
extensions==0.4
filelock==3.16.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.12.23
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.7.0
keybert==0.8.5
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgspec==0.19.0
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
namex==0.0.8
networkx==3.4.2
nlp==0.4.0
nltk==3.9.1
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
praw==7.8.1
prawcore==2.4.0
protobuf==5.29.2
pyarrow==18.1.0
Pygments==2.18.0
PyJWT==2.10.1
pymilvus==2.5.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
services==0.1.1
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
WTForms==3.2.1
xxhash==3.5.0
zope.interface==7.2
